Dev

Approved-by: Matthew Jones <mattyl.jones@dunelm.org.uk>

+++ morphct/utils/update_pickle.py
@@ -9,7 +9,7 @@ import numpy as np
 from morphct.code import helper_functions as hf
 from morphct.code import obtain_chromophores as oc
 from morphct.definitions import PROJECT_ROOT
-from morphct.templates import par_template
+from morphct.templates import par as par_template
 
 
 def convert_params(old_parameter_dict, file_name):
@@ -48,7 +48,9 @@ def add_underscores(old_parameter_dict):
         "DOS",
         "AAID",
     ]
-    leave_these_capitalised += [abbrev + "s" for abbrev in leave_these_capitalised]
+    leave_these_capitalised += [
+        "".join([abbrev, "s"]) for abbrev in leave_these_capitalised
+    ]
     new_parameter_dict = {}
     for key, value in old_parameter_dict.items():
         # Some hardcoded ones because I can't work out the regex
@@ -168,7 +170,11 @@ def rename_old(old_parameter_dict):
 
 
 def remove_redundant(old_parameter_dict):
-    remove_these_keys = ["electrical_field", "execute_extract_molecules"]
+    remove_these_keys = [
+        "electrical_field",
+        "execute_extract_molecules",
+        "koopmans_hopping_prefactor",
+    ]
     for key in remove_these_keys:
         try:
             old_parameter_dict.pop(key)
@@ -221,10 +227,9 @@ def convert_chromos(
     print("Updating the chromophore list (this might take a few minutes)...")
     for old_chromo in old_chromophore_list:
         print(
-            "\rUpdating chromophore",
-            old_chromo.ID + 1,
-            "of",
-            len(old_chromophore_list),
+            "\rUpdating chromophore {0:d} of {1:d}...".format(
+                old_chromo.ID + 1, len(old_chromophore_list)
+            ),
             end=" ",
         )
         # Set up an empty chromophore instance using the parameter_dict
@@ -348,6 +353,8 @@ def load_pickle_data(old_pickle_file, directory):
 
 def convert_KMC(morphology_directory):
     KMC_data = load_KMC_results_pickle(morphology_directory)
+    if not KMC_data:
+        return False
     print("Updating KMC result keys to PEP8 format...")
     KMC_data = add_underscores(KMC_data)
     print("Rewriting KMC results file...")
@@ -358,23 +365,26 @@ def convert_KMC(morphology_directory):
 
 
 def load_KMC_results_pickle(directory):
+    KMC_pickle = os.path.join(directory, "KMC", "KMCResults.pickle")
     try:
-        with open(directory + "/KMC/KMCResults.pickle", "rb") as pickle_file:
+        with open(KMC_pickle, "rb") as pickle_file:
             carrier_data = pickle.load(pickle_file)
     except FileNotFoundError:
         print("No final KMC_results.pickle found. Creating it from incomplete parts...")
-        create_results_pickle(directory)
-        with open(directory + "/KMC/KMC_results.pickle", "rb") as pickle_file:
+        success = create_results_pickle(directory)
+        if not success:
+            return False
+        with open(KMC_pickle, "rb") as pickle_file:
             carrier_data = pickle.load(pickle_file)
     except UnicodeDecodeError:
-        with open(directory + "/KMC/KMCResults.pickle", "rb") as pickle_file:
+        with open(KMC_pickle, "rb") as pickle_file:
             carrier_data = pickle.load(pickle_file, encoding="latin1")
     return carrier_data
 
 
 def create_results_pickle(directory):
     cores_list = []
-    for file_name in glob.glob(directory + "/KMC/*"):
+    for file_name in glob.glob(os.path.join(directory, "KMC", "*")):
         try:
             cores_list.append(re.findall("([_])(..)([\.])", file_name)[0][1])
         except IndexError:
@@ -384,14 +394,20 @@ def create_results_pickle(directory):
     keep_list = []
     for core in cores_list:
         # Check if there is already a finished KMC_results pickle
-        main = directory + "/KMC/KMCResults_%02d.pickle" % (int(core))
+        main = os.path.join(
+            directory, "KMC", "KMCResults_{:02d}.pickle".format(int(core))
+        )
         if os.path.exists(main):
             results_pickles_list.append(main)
             keep_list.append(None)
             continue
         # If not, find the slot1 and slot2 pickle that is most recent
-        slot1 = directory + "/KMC/KMCSlot1Results_%02d.pickle" % (int(core))
-        slot2 = directory + "/KMC/KMCSlot2Results_%02d.pickle" % (int(core))
+        slot1 = os.path.join(
+            directory, "KMC", "KMCSlot1Results_{:02d}.pickle".format(int(core))
+        )
+        slot2 = os.path.join(
+            directory, "KMC", "KMCSlot2Results_{:02d}.pickle".format(int(core))
+        )
         if os.path.exists(slot1) and not os.path.exists(slot2):
             keep_list.append(slot1)
         elif os.path.exists(slot2) and not os.path.exists(slot1):
@@ -400,13 +416,18 @@ def create_results_pickle(directory):
             keep_list.append(slot1)
         else:
             keep_list.append(slot2)
-    print("%d pickle files found to combine!" % (len(keep_list)))
+    print("{:d} pickle files found to combine!".format(len(keep_list)))
+    if len(keep_list) == 0:
+        print("No KMC data found. Skipping...")
+        return False
     print("Combining", keep_list)
     for keeper in zip(cores_list, keep_list):
         # Skip this core if we already have a finished KMC_results for it
         if keeper[1] is None:
             continue
-        new_name = directory + "/KMC/KMC_results_" + str(keeper[0]) + ".pickle"
+        new_name = os.path.join(
+            directory, "KMC", "KMC_results_{}.pickle".format(keeper[0])
+        )
         shutil.copyfile(str(keeper[1]), new_name)
         results_pickles_list.append(new_name)
     combine_results_pickles(directory, results_pickles_list)
@@ -429,9 +450,10 @@ def combine_results_pickles(directory, pickle_files):
                     combined_data[key] += val
     # Write out the combined data
     print("Writing out the combined pickle file...")
-    with open(directory + "/KMC/KMC_results.pickle", "wb+") as pickle_file:
+    combined_file_loc = os.path.join(directory, "KMC", "KMC_results.pickle")
+    with open(combined_file_loc, "wb+") as pickle_file:
         pickle.dump(combined_data, pickle_file)
-    print("Complete data written to", directory + "/KMC_results.pickle.")
+    print("Complete data written to", combined_file_loc)
 
 
 def main():
@@ -458,6 +480,7 @@ def main():
         print("Update_pickle is running in `directory mode' (default).")
     # Iterate over all input strings (directories or pickles)
     for input_string in input_list:
+        print()
         if args.pickle_mode:
             # Load the pickle directly, and do not output a parameter file
             # NOTE: It looks like old_ and new_ are the wrong way around here
